Fix some compiler warnings

+++ IRCCloud/Classes/MainViewController.m
@@ -1521,7 +1521,7 @@ extern NSDictionary *emojiMap;
             if(diff > 0 || _buffer.scrolledUp)
                 _eventsView.tableView.contentOffset = CGPointMake(0, _eventsView.tableView.contentOffset.y + diff);
             else if([[UIDevice currentDevice].systemVersion isEqualToString:@"8.1"]) //The last line seems to get eaten when the table is fully scrolled down on iOS 8.1
-                _eventsView.tableView.contentOffset = CGPointMake(0, _eventsView.tableView.contentOffset.y + fabsf(diff));
+                _eventsView.tableView.contentOffset = CGPointMake(0, _eventsView.tableView.contentOffset.y + fabs(diff));
         }
     }
 }
